requirements.txt
torch
torchvision
torchaudio
transformers
datasets
sentencepiece
openai-whisper
TTS
tensorflow
langchain
langchain-community
faiss-cpu
sentence-transformers
gradio
wandb
pandas
numpy
matplotlib
pillow
reportlab
pypdf
python-dotenv
requests